fix(pointers): Fix pointer velocities are invalid when manipulation interia is running
